chore: better support of typescript
